pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clawler"
version = "5.0.0"
description = "Advanced news crawling service — no API keys required"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "Clawdia @ OpenClaw"}]
keywords = ["news", "rss", "crawler", "aggregator", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.0",
    "python-dateutil>=2.8.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
clawler = "clawler.cli:main"

[project.urls]
Repository = "https://github.com/clawdiard/clawler"

[tool.pytest.ini_options]
testpaths = ["tests"]
